Changed main.cpp

+++ main.cpp
@@ -3,5 +3,5 @@
 int main(int argc, char* argv[]) {
    App app(argc, argv);
 
-   return app.run();
+   return (int)app.run();
 }
